+++ src/Spice86.Core/Emulator/CPU/CfgCpu/ParsedInstruction/Instructions/PushF16.cs
@@ -5,7 +5,7 @@ using Spice86.Core.Emulator.CPU.CfgCpu.ParsedInstruction.Prefix;
 using Spice86.Shared.Emulator.Memory;
 
 public class PushF16 : CfgInstruction {
-    public PushF16(SegmentedAddress address, InstructionField<byte> opcodeField, List<InstructionPrefix> prefixes) :
+    public PushF16(SegmentedAddress address, InstructionField<ushort> opcodeField, List<InstructionPrefix> prefixes) :
         base(address, opcodeField, prefixes) {
     }
